Fixed bug: Wrong dtype in cylinder rendering

+++ src/molara/Rendering/rendering.py
@@ -94,10 +94,10 @@ class Renderer:
 
         for i in range(n_instances):
             model_matrix = calculate_cylinder_model_matrix(
-                np.array(positions[i], dtype=np.float64),
+                np.array(positions[i], dtype=np.float32),
                 float(radii[i]),
                 float(lengths[i]),
-                np.array(directions[i], dtype=np.float64),
+                np.array(directions[i], dtype=np.float32),
             )
             model_matrices = model_matrix if i == 0 else np.concatenate((model_matrices, model_matrix))
